Add files via upload
The project involves applying a machine learning clustering algorithm to the country_data.csv dataset, which contains information about various countries' child mortality rates, exports, health spending, and other relevant features. The objective is to explore the dataset using clustering techniques, visualize the resulting clusters, and analyze the findings. The analysis focus on drawing conclusions based on the clusters, considering factors such as the inclusion or exclusion of specific features and their impact on the clustering results. Additionally, the project aims to provide insights and recommendations within the context of the data, based on the clustering outcomes.